+++ modules/assim.sequential/inst/WillowCreek/SDA_Workflow.R
@@ -563,12 +563,12 @@ source('/fs/data3/kzarada/pecan/modules/assim.sequential/R/Nimble_codes.R')
 
 
 if(restart == FALSE) unlink(c('run','out','SDA'), recursive = T)
-debugonce(PEcAn.assim.sequential::sda.enkf)
+debugonce(PEcAnAssimSequential::sda.enkf)
 
 if ('state.data.assimilation' %in% names(settings)) {
   if (PEcAn.utils::status.check("SDA") == 0) {
     PEcAn.utils::status.start("SDA")
-    PEcAn.assim.sequential::sda.enkf(
+    PEcAnAssimSequential::sda.enkf(
       settings, 
       restart=restart,
       Q=0,